Moved to "mistake". This requires more thought.
At issue is that the "recently" made validation types "wchan" and "rwchan" generate externally visible changes as part of their creation of the internal representation. Namely, they create a file. This is problematic as in some situations the VT is called on to simply check if the input is ok, and nothing more. And while the code path in question does call on the "release" method of the VT to cleanup the internal rep the external side-effect remains.

The change here went with extending "release" to allow it to undo the external effect as well, on request. This then requires determining when to do that, and when not. And that is difficult as the same APIs are used internally and given to the user.

A better path might be to change the VT API. For compatibility we cannot change "validate", it has to keep doing validation + conversion. New methods would be required for each of the two parts. Then the framework can defer the creation of the int. rep until the last possible moment, i.e. when the user needs it. While the other places only requiring validation limit themselves to that, and no side-effects, in- or outside of memory.

----
Extended the std validation types to support an undo flag for their release method.