cmdr::actor, cmdr::officer - Fix bug introduced with revision [7ab77ead21].
The option handling added to "actor::set" in that revision means that
the command "my set *command* ..." storing the actual command line
(i.e. user information) can break, trying to interpret application
specific option information as something for the internal command.

Fixed by adding "--" to "actor::set"s option handling, and using it
for the breakable command to force interpretation of the user data as
such.