cmdr::config - Fix long-standing bug in forced parameter value calculation (method 'Force').
Any error in the calculations left the flag 'myinforce' set, causing a
future invokation to believe to be in a recursive call and do
nothing. While this had no effect on regular operation, i.e. where the
using application exits after the command, in interactive mode this
disabled checks and validations for the command in question, and also
retained old parameter values. Fixed by putting a try/finally around
the section, resetting the flag even in the presence of errors thrown
by it.

cmdr::config version bumped to 1.1.1.